增加一个编解码器

+++ mmorpg-server/src/main/java/com/common/pack/RequestPacket.java
@@ -13,7 +13,7 @@ public class RequestPacket {
         return rp;
     }
 
-    public int getPacketId() {
+    public short getPacketId() {
         return packetId;
     }
